Put GUI elements back into main.py to avoid opening two display windows. Deleted gui.py
